requirements.txt
requests>=2.31.0
groq>=0.4.0
python-dotenv>=1.0.0
streamlit>=1.31.0
pandas>=2.0.0
praw>=7.7.0
openai>=1.0.0
